+++ components/src/lib.rs
@@ -9,6 +9,7 @@ mod drag_drop;
 mod dropdown;
 mod external_link;
 mod gesture_area;
+mod graph;
 mod input;
 mod loader;
 mod network_image;
@@ -26,6 +27,7 @@ pub use drag_drop::*;
 pub use dropdown::*;
 pub use external_link::*;
 pub use gesture_area::*;
+pub use graph::*;
 pub use input::*;
 pub use loader::*;
 pub use network_image::*;
